requirements.txt
google-adk
google-genai
google-generativeai
python-dotenv
speechrecognition
pyaudio
ddgs
requests
beautifulsoup4
trafilatura
gTTS
edge-tts
pygame
pyautogui
customtkinter
youtube-search-python
python-docx
openai
